requirements.txt
attrs
numpy
pandas
plotnine
plydata
pytest
requests
scikit-learn
scikit-surprise
tqdm
